+++ packages/react-native-sdk/src/components/participants/LocalParticipantView.tsx
@@ -17,6 +17,7 @@ import { Avatar } from '../utility/Avatar';
 import { LOCAL_VIDEO_VIEW_STYLE, Z_INDEX } from '../../constants';
 import { useDebouncedValue } from '../../utils/hooks';
 import { useMediaStreamManagement } from '../../providers/MediaStreamManagement';
+import { ParticipantReaction } from './internal/ParticipantReaction';
 
 /**
  * Props to be passed for the LocalVideoView component.
@@ -101,17 +102,24 @@ export const LocalParticipantView = (props: LocalParticipantViewProps) => {
   );
 
   if (layout === 'fullscreen') {
-    if (isVideoMuted) {
-      return (
-        <View
-          accessibilityLabel={A11yComponents.LOCAL_PARTICIPANT_FULLSCREEN}
-          style={styles.avatarContainer}
-        >
-          <Avatar participant={localParticipant} />
+    return (
+      <View
+        accessibilityLabel={A11yComponents.LOCAL_PARTICIPANT_FULLSCREEN}
+        style={style}
+      >
+        <View style={styles.topView}>
+          <ParticipantReaction
+            reaction={localParticipant.reaction}
+            sessionId={localParticipant.sessionId}
+          />
         </View>
-      );
-    }
-    return VideoRenderOrBlankComponent;
+        {isVideoMuted ? (
+          <Avatar participant={localParticipant} />
+        ) : (
+          VideoRenderOrBlankComponent
+        )}
+      </View>
+    );
   }
 
   return (
@@ -124,15 +132,21 @@ export const LocalParticipantView = (props: LocalParticipantViewProps) => {
       }}
       {...panResponder.panHandlers}
     >
-      {isVideoMuted ? (
-        <View style={style}>
+      <View style={style}>
+        <View style={styles.topView}>
+          <ParticipantReaction
+            reaction={localParticipant.reaction}
+            sessionId={localParticipant.sessionId}
+          />
+        </View>
+        {isVideoMuted ? (
           <View style={theme.icon.md}>
             <VideoSlash color={theme.light.static_white} />
           </View>
-        </View>
-      ) : (
-        VideoRenderOrBlankComponent
-      )}
+        ) : (
+          VideoRenderOrBlankComponent
+        )}
+      </View>
     </Animated.View>
   );
 };
@@ -146,10 +160,24 @@ const styles = StyleSheet.create({
     top: theme.margin.xl * 2,
     borderRadius: LOCAL_VIDEO_VIEW_STYLE.borderRadius,
     zIndex: Z_INDEX.IN_MIDDLE,
-    overflow: 'hidden',
-    backgroundColor: theme.light.disabled,
+    backgroundColor: theme.light.static_grey,
     justifyContent: 'center',
     alignItems: 'center',
+    overflow: 'hidden',
+  },
+  videoStream: {
+    height: LOCAL_VIDEO_VIEW_STYLE.height,
+    width: LOCAL_VIDEO_VIEW_STYLE.width,
+    flex: 1,
+  },
+  videoStreamFullScreen: {
+    ...StyleSheet.absoluteFillObject,
+  },
+  topView: {
+    position: 'absolute',
+    top: theme.spacing.sm,
+    left: theme.spacing.sm,
+    zIndex: Z_INDEX.IN_FRONT,
   },
   fullScreenContainer: {
     flex: 1,