+++ sample-apps/react/audio-rooms/src/components/Room/RoomUI.tsx
@@ -2,16 +2,20 @@ import {
   CallingState,
   OwnCapability,
   Restricted,
+  SfuEvents,
+  SfuModels,
+  StreamVideoEvent,
   StreamVideoLocalParticipant,
   StreamVideoParticipant,
   useCall,
   useCallCallingState,
   useCallMetadata,
   useHasPermissions,
+  useLocalParticipant,
   useParticipants,
 } from '@stream-io/video-react-sdk';
-import { useMemo } from 'react';
-import { ChatIcon, PersonIcon } from '../icons';
+import { useCallback, useEffect, useMemo, useState } from 'react';
+import { ChatIcon, CloseIcon, ListIcon, PersonIcon } from '../icons';
 import { SpeakerElement } from './SpeakerElement';
 import { SpeakingRequestsList } from './SpeakingRequestsList';
 import { useSpeakingRequests } from '../../hooks/useSpeakingRequests';
@@ -20,14 +24,19 @@ import { LiveRoomControls } from './LiveRoomControls';
 import { EndedRoomOverlay, RoomLobby } from './Overlay';
 import { RoomListing } from '../RoomList';
 import { RoomAccessControls } from './RoomAccessControls';
-import { useLayoutController } from '../../contexts';
 import type { CustomCallData } from '../../types';
+import { RoomListingTabs } from '../RoomList/RoomListingTabs';
+import { RoomLiveState } from '../../utils/roomLiveState';
 
-export const RoomUI = () => {
-  const { showRoomList } = useLayoutController();
+type RoomUIProps = {
+  loadRoom: () => Promise<void>;
+};
+
+export const RoomUI = ({ loadRoom }: RoomUIProps) => {
   const call = useCall();
   const callMetadata = useCallMetadata();
   const callingState = useCallCallingState();
+  const localParticipant = useLocalParticipant();
   const participants = useParticipants();
   const canJoinBackstage = useHasPermissions(OwnCapability.JOIN_BACKSTAGE);
   const {
@@ -41,6 +50,46 @@ export const RoomUI = () => {
     hosts = [],
     speakerIds = [],
   } = (callMetadata?.custom || {}) as CustomCallData;
+
+  const [showRoomList, setShowRoomList] = useState(false);
+
+  const toggleShowRoomList = useCallback(
+    () => setShowRoomList((prev) => !prev),
+    [],
+  );
+
+  useEffect(() => {
+    if (!call || !localParticipant) return;
+
+    const unsubscribeFromLiveEnded = call.on(
+      'error',
+      (e: SfuEvents.SfuEvent) => {
+        if (
+          e.eventPayload.oneofKind !== 'error' ||
+          !e.eventPayload.error.error ||
+          e.eventPayload.error.error.code !== SfuModels.ErrorCode.LIVE_ENDED
+        )
+          return;
+        if (
+          !call.permissionsContext.hasPermission(OwnCapability.JOIN_BACKSTAGE)
+        )
+          loadRoom();
+      },
+    );
+
+    const unsubscribeFromParticipantLeft = call.on(
+      'call.session_participant_left',
+      (e: StreamVideoEvent) => {
+        if (e.type !== 'call.session_participant_left') return;
+
+        if (e.user_session_id === localParticipant.sessionId) loadRoom();
+      },
+    );
+    return () => {
+      unsubscribeFromLiveEnded();
+      unsubscribeFromParticipantLeft();
+    };
+  }, [loadRoom, call, localParticipant]);
 
   const { speakers, listeners } = useMemo(() => {
     const hostIds = hosts.map((host) => host.id) || [];
@@ -70,57 +119,80 @@ export const RoomUI = () => {
   const isRoomEnded = !!callMetadata?.ended_at;
 
   return (
-    <section className="active-room">
-      <section className={`rooms-overview ${!showRoomList ? 'hidden' : ''}`}>
-        <RoomListing liveState="live" />
-        <RoomListing liveState="upcoming" />
-        <RoomListing liveState="ended" />
-      </section>
-      <div className={`room-detail ${showRoomList ? 'with-room-list' : ''}`}>
-        <div className="room-detail-header">
-          <h2>{title}</h2>
-          <LiveRoomControls
-            hasNotifications={speakingRequests.length > 0}
-            openRequestsList={() => setIsOpenRequestList(true)}
-          />
-        </div>
-        <p className="user-counts secondaryText">
-          {participants.length}
-          <PersonIcon />/ {speakers.length}
-          <ChatIcon />
-        </p>
-        <section className="participants-section">
-          <h3>Speakers ({speakers.length})</h3>
-          <div className="speakers-list">
-            {speakers.map((speaker) => (
-              <SpeakerElement key={speaker.sessionId} speaker={speaker} />
-            ))}
-          </div>
-        </section>
-        <section className="participants-section">
-          <h3>Listeners ({listeners.length})</h3>
-          <div className="listeners-list">
-            {listeners.map((listener) => (
-              <Listener key={listener.sessionId} participant={listener} />
-            ))}
-          </div>
-        </section>
-        <RoomAccessControls />
-        {isOpenRequestList && (
-          <Restricted
-            requiredGrants={[OwnCapability.UPDATE_CALL_PERMISSIONS]}
-            hasPermissionsOnly
-          >
-            <SpeakingRequestsList
-              close={() => setIsOpenRequestList(false)}
-              dismissSpeakingRequest={dismissSpeakingRequest}
-              speakingRequests={speakingRequests}
+    <>
+      <button
+        className="show-room-list-button filled-button filled-button--blue"
+        onClick={toggleShowRoomList}
+        title={`${showRoomList ? 'Hide' : 'Show'} rooms`}
+      >
+        {showRoomList ? <CloseIcon /> : <ListIcon />}
+        {/*<span>{`${showRoomList ? 'Hide' : 'Show'} rooms`}</span>*/}
+      </button>
+
+      <section
+        className={`active-room ${showRoomList ? 'with-room-list' : ''}`}
+      >
+        {showRoomList && <RoomsListing />}
+        <div className={`room-detail`}>
+          <div className="room-detail-header">
+            <h2>{title}</h2>
+            <LiveRoomControls
+              hasNotifications={speakingRequests.length > 0}
+              openRequestsList={() => setIsOpenRequestList(true)}
             />
-          </Restricted>
-        )}
-        {isRoomEnded && <EndedRoomOverlay />}
-        {showLobby && <RoomLobby />}
-      </div>
+          </div>
+          <p className="user-counts secondaryText">
+            {participants.length}
+            <PersonIcon />/ {speakers.length}
+            <ChatIcon />
+          </p>
+          <section className="participants-section">
+            <h3>Speakers ({speakers.length})</h3>
+            <div className="speakers-list">
+              {speakers.map((speaker) => (
+                <SpeakerElement key={speaker.sessionId} speaker={speaker} />
+              ))}
+            </div>
+          </section>
+          <section className="participants-section">
+            <h3>Listeners ({listeners.length})</h3>
+            <div className="listeners-list">
+              {listeners.map((listener) => (
+                <Listener key={listener.sessionId} participant={listener} />
+              ))}
+            </div>
+          </section>
+          <RoomAccessControls />
+          {isOpenRequestList && (
+            <Restricted
+              requiredGrants={[OwnCapability.UPDATE_CALL_PERMISSIONS]}
+              hasPermissionsOnly
+            >
+              <SpeakingRequestsList
+                close={() => setIsOpenRequestList(false)}
+                dismissSpeakingRequest={dismissSpeakingRequest}
+                speakingRequests={speakingRequests}
+              />
+            </Restricted>
+          )}
+          {isRoomEnded && <EndedRoomOverlay />}
+          {showLobby && <RoomLobby />}
+        </div>
+      </section>
+    </>
+  );
+};
+
+const RoomsListing = () => {
+  const [activeLiveState, setActiveLiveState] = useState<RoomLiveState>('live');
+
+  return (
+    <section className={`rooms-overview`}>
+      <RoomListingTabs
+        activeLiveState={activeLiveState}
+        onSelect={setActiveLiveState}
+      />
+      <RoomListing liveState={activeLiveState} />
     </section>
   );
 };