+++ packages/react-native-sdk/CHANGELOG.md
@@ -2,6 +2,32 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.0.1-alpha.263](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.262...@stream-io/video-react-native-sdk-0.0.1-alpha.263) (2023-07-10)
+
+
+### Bug Fixes
+
+* **react-native:** video published on foreground even if it was muted ([#773](https://github.com/GetStream/stream-video-js/issues/773)) ([6fc1ce9](https://github.com/GetStream/stream-video-js/commit/6fc1ce91db183e6fa2c962b86b5715e3ac4421c6))
+
+
+
+## [0.0.1-alpha.262](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.261...@stream-io/video-react-native-sdk-0.0.1-alpha.262) (2023-07-10)
+
+
+### Bug Fixes
+
+* **react-native:** fix ui issues in dogfood app ([#763](https://github.com/GetStream/stream-video-js/issues/763)) ([289374c](https://github.com/GetStream/stream-video-js/commit/289374c0d17f7897906fab22ee2e06be0380eb51))
+
+
+
+## [0.0.1-alpha.261](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.260...@stream-io/video-react-native-sdk-0.0.1-alpha.261) (2023-07-10)
+
+
+
+## [0.0.1-alpha.260](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.259...@stream-io/video-react-native-sdk-0.0.1-alpha.260) (2023-07-10)
+
+
+
 ## [0.0.1-alpha.259](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.258...@stream-io/video-react-native-sdk-0.0.1-alpha.259) (2023-07-07)